% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estructura.R
\name{area_basal}
\alias{area_basal}
\title{Área Basal (§6.8 MINAM 2015)}
\usage{
area_basal(dap_cm = NULL, lc_cm = NULL, unidad_salida = c("m2", "cm2"))
}
\arguments{
\item{dap_cm}{Vector numérico con el DAP en centímetros.}

\item{lc_cm}{Vector numérico opcional con la longitud de circunferencia en cm.}

\item{unidad_salida}{Carácter; \code{"m2"} (metros cuadrados) o \code{"cm2"}. Por defecto \code{"m2"}.}
}
\value{
Vector numérico con el área basal en la unidad especificada.
}
\description{
Calcula el área basal individual o agregada a partir del Diámetro a la Altura del Pecho (DAP en cm)
o de la Longitud de Circunferencia (LC en cm).
}
\examples{
area_basal(dap_cm = c(15, 25, 40))
}
