% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_t}
\alias{bci_t}
\title{Datos de Inventario de Arboles en la Isla Barro Colorado (BCI Dataset en Formato Largo)}
\format{
Un \code{data.frame} con 4539 filas y 3 columnas:
\describe{
\item{plot}{Identificador de la parcela / sitio de 1 ha (1 al 50).}
\item{species}{Nombre cientifico completo actualizado de la especie arborea (225 especies).}
\item{abundance}{Numero de individuos registrados para la especie en la parcela.}
}
}
\source{
Zanne, A.E. et al. (2014). Three keys to the terrestrial biodiversity of Barro Colorado Island. \emph{Dryad Digital Repository}.
}
\usage{
data(bci_t)
}
\description{
Registro de inventario de arboles en 50 parcelas de 1 hectarea en la Isla Barro Colorado (BCI, Panama;
Zanne et al., 2014; Condit et al., 2002) con 225 especies arboreas, formateado en estructura larga
(sitio/plot, especie/species, abundancia/abundance).
}
\keyword{datasets}
