% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{curva_acumulacion}
\alias{curva_acumulacion}
\title{Curva de Acumulación de Especies (MINAM 2015)}
\usage{
curva_acumulacion(
  datos,
  sitio = "sitio",
  especie = "especie",
  abundancia = "abundancia",
  metodo = "random",
  sustituciones = 100
)
}
\arguments{
\item{datos}{Data frame en formato long o matriz de comunidad.}

\item{sitio}{Nombre de la columna del sitio. Por defecto \code{"sitio"}.}

\item{especie}{Nombre de la columna de la especie. Por defecto \code{"especie"}.}

\item{abundancia}{Nombre de la columna de abundancia. Por defecto \code{"abundancia"}.}

\item{metodo}{Método de acumulación para \code{vegan::specaccum}. Por defecto \code{"random"}.}

\item{sustituciones}{Número de permutaciones cuando \code{metodo = "random"}. Por defecto \code{100}.}
}
\value{
Objeto de clase \code{specaccum}.
}
\description{
Curva de Acumulación de Especies (MINAM 2015)
}
