% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estructura.R
\name{densidad_poblacional}
\alias{densidad_poblacional}
\title{Densidad Poblacional (§6.4 MINAM 2015)}
\usage{
densidad_poblacional(n_individuos, area_ha, tocones = 0)
}
\arguments{
\item{n_individuos}{Número total o vector de individuos.}

\item{area_ha}{Área del muestreo en hectáreas (ha).}

\item{tocones}{Conteo adicional de tocones. Por defecto \code{0}.}
}
\value{
Densidad en individuos/ha.
}
\description{
Calcula la densidad poblacional (individuos por hectárea), considerando opcionalmente conteo de tocones.
}
\examples{
densidad_poblacional(n_individuos = 450, area_ha = 0.5)
}
