% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasa_volumen.R
\name{densidades_referencia_neotropical}
\alias{densidades_referencia_neotropical}
\title{Base de Datos de Referencia Neotropical para Densidad Básica de la Madera (g/cm³)}
\usage{
densidades_referencia_neotropical
}
\value{
Vector numérico nombrado con valores de densidad básica de la madera (g/cm3).
}
\description{
Base de Datos de Referencia Neotropical para Densidad Básica de la Madera (g/cm³)
}
\examples{
head(densidades_referencia_neotropical)
}
