% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{diversidad_pielou}
\alias{diversidad_pielou}
\title{Índice de Equitabilidad de Pielou (J') (SINIA / MINAM Tabla 2.0-4)}
\usage{
diversidad_pielou(
  datos,
  sitio = "sitio",
  especie = "especie",
  abundancia = "abundancia",
  base = 2
)
}
\arguments{
\item{datos}{Data frame en formato long o matriz de comunidad.}

\item{sitio}{Nombre de la columna del sitio. Por defecto \code{"sitio"}.}

\item{especie}{Nombre de la columna de la especie. Por defecto \code{"especie"}.}

\item{abundancia}{Nombre de la columna de abundancia. Por defecto \code{"abundancia"}.}

\item{base}{Base logarítmica utilizada en Shannon (2 por defecto).}
}
\value{
Data frame con el sitio, riqueza (S), Shannon (H') y Equitabilidad de Pielou (J').
}
\description{
Índice de Equitabilidad de Pielou (J') (SINIA / MINAM Tabla 2.0-4)
}
