% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{diversidad_shannon}
\alias{diversidad_shannon}
\title{Índice de Diversidad de Shannon-Wiener (H') (MINAM 2015)}
\usage{
diversidad_shannon(
  datos,
  sitio = "sitio",
  especie = "especie",
  abundancia = "abundancia",
  base = 2
)
}
\arguments{
\item{datos}{Data frame en formato long o matriz de comunidad.}

\item{sitio}{Nombre de la columna del sitio. Por defecto \code{"sitio"}.}

\item{especie}{Nombre de la columna de la especie. Por defecto \code{"especie"}.}

\item{abundancia}{Nombre de la columna de abundancia. Por defecto \code{"abundancia"}.}

\item{base}{Base del logaritmo. Por defecto \code{2}.}
}
\value{
Data frame con el sitio y su respectivo valor de Shannon-Wiener (H').
}
\description{
Índice de Diversidad de Shannon-Wiener (H') (MINAM 2015)
}
