% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{diversidad_simpson}
\alias{diversidad_simpson}
\title{Índice de Dominancia y Diversidad de Simpson (D y 1-D)}
\usage{
diversidad_simpson(
  datos,
  sitio = "sitio",
  especie = "especie",
  abundancia = "abundancia"
)
}
\arguments{
\item{datos}{Data frame en formato long o matriz de comunidad.}

\item{sitio}{Nombre de la columna del sitio. Por defecto \code{"sitio"}.}

\item{especie}{Nombre de la columna de la especie. Por defecto \code{"especie"}.}

\item{abundancia}{Nombre de la columna de abundancia. Por defecto \code{"abundancia"}.}
}
\value{
Data frame con el sitio, dominancia de Simpson (D) e índice de Simpson (1 - D).
}
\description{
Índice de Dominancia y Diversidad de Simpson (D y 1-D)
}
