% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{diversidad_sorensen}
\alias{diversidad_sorensen}
\title{Coeficiente de Similitud de Sorensen-Dice}
\usage{
diversidad_sorensen(
  datos,
  sitio = "sitio",
  especie = "especie",
  abundancia = "abundancia"
)
}
\arguments{
\item{datos}{Data frame en formato long o matriz de comunidad.}

\item{sitio}{Nombre de la columna del sitio. Por defecto \code{"sitio"}.}

\item{especie}{Nombre de la columna de la especie. Por defecto \code{"especie"}.}

\item{abundancia}{Nombre de la columna de abundancia. Por defecto \code{"abundancia"}.}
}
\value{
Objeto de clase \code{dist} con los valores de similitud de Sorensen-Dice.
}
\description{
Coeficiente de Similitud de Sorensen-Dice
}
