% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dune_t}
\alias{dune_t}
\title{Datos de Observaciones de Vegetacion de Dunas (Dune Dataset en Formato Largo)}
\format{
Un \code{data.frame} con 197 filas y 3 columnas:
\describe{
\item{plot}{Identificador de la parcela / sitio (1 al 20).}
\item{species}{Nombre abreviado o cientifico de la especie (30 especies).}
\item{abundance}{Conteo / abundancia observada de la especie en el sitio.}
}
}
\source{
Jongman, R.H.G., ter Braak, C.J.F. & van Tongeren, O.F.R. (1987). \emph{Data Analysis in Community and Landscape Ecology}. Pudoc, Wageningen.
}
\usage{
data(dune_t)
}
\description{
Conjunto de datos de observaciones de 30 especies de vegetacion en 20 sitios en dunas holandesas
(Jongman et al., 1987), formateado en estructura larga (sitio/plot, especie/species, abundancia/abundance).
}
\keyword{datasets}
