% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floraveg-package.R
\docType{package}
\name{floraveg-package}
\alias{floraveg-package}
\alias{floraveg}
\title{floraveg: Indicadores Ecológicos y de Vegetación según Guía MINAM}
\description{
Paquete en R para automatizar el cálculo de parámetros estructurales,
índices de diversidad, biomasa, volumen maderable e Índice de Valor de Importancia (IVI),
conforme a la Guía de Inventario de la Flora y Vegetación (MINAM, 2015).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/floraveg}
  \item \url{https://paulesantos.github.io/floraveg/}
  \item Report bugs at \url{https://github.com/PaulESantos/floraveg/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

}
\keyword{internal}
