% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasa_volumen.R
\name{obtener_densidad_madera}
\alias{obtener_densidad_madera}
\title{Obtener Densidades Específicas de la Madera por Especie o Género}
\usage{
obtener_densidad_madera(especies, db_densidades = NULL, default = 0.6)
}
\arguments{
\item{especies}{Vector de caracteres con nombres científicos o identificadores de especie.}

\item{db_densidades}{Vector nombrado opcional con valores de densidad (g/cm3). Si es \code{NULL},
se utiliza la base de referencia neotropical interna.}

\item{default}{Valor numérico por defecto (g/cm3) para especies o géneros no encontrados. Por defecto \code{0.60}.}
}
\value{
Un vector numérico de densidades básicas de la madera de la misma longitud que \code{especies}.
}
\description{
Asigna valores de densidad básica de la madera (g/cm3) consultando un diccionario
de referencia por especie (ej. 'Cedrela odorata'), por género (ej. 'Cedrela') o retornando
un valor estándar predeterminado en caso de no encontrarse coincidencia.
}
\examples{
especies <- c("Cedrela odorata", "Inga edulis", "Especie desconocida")
obtener_densidad_madera(especies, default = 0.60)
}
