% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversidad.R
\name{plot_whittaker}
\alias{plot_whittaker}
\title{Grafico ggplot2 de Curvas de Dominancia-Abundancia de Whittaker}
\usage{
plot_whittaker(datos, scale = "logabun", por_sitio = TRUE, top_n_labels = 3)
}
\arguments{
\item{datos}{Data frame con datos de inventario.}

\item{scale}{Escala de abundancia en el eje Y: \code{"logabun"} (Log10 Abundancia), \code{"abundance"} (Absoluta), \code{"proportion"} (porcentaje Abundancia Relativa), \code{"accumfreq"} (porcentaje Frecuencia Acumulada). Por defecto \code{"logabun"}.}

\item{por_sitio}{Si es \code{TRUE}, grafica curvas por sitio; si es \code{FALSE}, agrupa la comunidad general. Por defecto \code{TRUE}.}

\item{top_n_labels}{Numero de especies top a etiquetar en el grafico. Por defecto \code{3}.}
}
\value{
Un objeto de clase \code{ggplot}.
}
\description{
Genera un grafico de ggplot2 para visualizar distribuciones de abundancia de especies (SAD).
}
