% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_floraveg.R
\name{run_floraveg}
\alias{run_floraveg}
\title{Lanzador de la Aplicación Shiny de floraveg}
\usage{
run_floraveg(launch.browser = TRUE, port = getOption("shiny.port"))
}
\arguments{
\item{launch.browser}{Lógico; si es \code{TRUE}, abre automáticamente el navegador. Por defecto \code{TRUE}.}

\item{port}{Puerto numérico opcional para el servidor Shiny.}
}
\value{
Objeto de aplicación Shiny (ejecución interactiva).
}
\description{
Inicia la interfaz gráfica interactiva del paquete \code{floraveg} en el navegador web predeterminado.
}
\examples{
\dontrun{
run_floraveg()
}
}
