% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modules.R
\name{run_modules}
\alias{run_modules}
\alias{run_mod_modelo_datos}
\alias{run_mod_diversidad}
\alias{run_mod_estructura}
\alias{run_mod_biomasa}
\alias{run_mod_ivi}
\alias{run_mod_codigo_r}
\title{Lanzadores Standalone de Módulos Shiny de floraveg}
\usage{
run_mod_modelo_datos(launch.browser = TRUE)

run_mod_diversidad(launch.browser = TRUE)

run_mod_estructura(launch.browser = TRUE)

run_mod_biomasa(launch.browser = TRUE)

run_mod_ivi(launch.browser = TRUE)

run_mod_codigo_r(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Lógico; abre en el navegador. Por defecto TRUE.}
}
\value{
Objeto Shiny app (ejecución interactiva).
}
\description{
Estas funciones permiten ejecutar cada módulo de la aplicación Shiny de forma 100\% independiente.
}
\examples{
if (interactive()) {
  run_mod_modelo_datos()
  run_mod_diversidad()
  run_mod_estructura()
  run_mod_biomasa()
  run_mod_ivi()
  run_mod_codigo_r()
}
}
