% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix.R
\name{validate_inventario}
\alias{validate_inventario}
\title{Validar estructura de datos de inventario}
\usage{
validate_inventario(datos, required_cols = c("sitio", "especie"))
}
\arguments{
\item{datos}{Data.frame o tibble con los datos del inventario.}

\item{required_cols}{Vector de caracteres con los nombres de las columnas indispensables.}
}
\value{
El mismo data frame si la validación es exitosa.
}
\description{
Comprueba que el objeto provisto sea un data frame y contenga las columnas requeridas.
}
\examples{
df <- data.frame(sitio = c("P1", "P1"), especie = c("Sp1", "Sp2"), abundancia = c(5, 2))
validate_inventario(df, c("sitio", "especie"))
}
