% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasa_volumen.R
\name{densidades_referencia_neotropical}
\alias{densidades_referencia_neotropical}
\title{Base de Datos de Referencia Neotropical para Densidad Básica de la Madera (g/cm³)}
\usage{
densidades_referencia_neotropical
}
\description{
Base de Datos de Referencia Neotropical para Densidad Básica de la Madera (g/cm³)
}
